requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
praw==7.4.0
prawcore==2.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.4
websocket-client==0.59.0
yarl==1.6.3
